+++ apps/forms/app/(api)/v1/[id]/route.ts
@@ -16,7 +16,6 @@ import {
   SYSTEM_GF_CUSTOMER_EMAIL_KEY,
   SYSTEM_GF_CUSTOMER_UUID_KEY,
   SYSTEM_GF_FINGERPRINT_VISITORID_KEY,
-  SYSTEM_GF_KEY_STARTS_WITH,
 } from "@/k/system";
 import { FormBlockTree } from "@/lib/forms/types";
 import { client } from "@/lib/supabase/server";
@@ -30,7 +29,6 @@ import {
   validate_max_access_by_customer,
   validate_max_access_by_form,
 } from "@/services/form/validate-max-access";
-import { is_uuid_v4 } from "@/utils/is";
 import i18next from "i18next";
 import { cookies } from "next/headers";
 import { notFound } from "next/navigation";
@@ -43,6 +41,7 @@ import type {
 } from "@/types";
 import { Features } from "@/lib/features/scheduling";
 import { requesturl } from "@/services/form/session-storage";
+import { type GFKeys, parseGFKeys } from "@/lib/forms/gfkeys";
 
 export const revalidate = 0;
 
@@ -137,9 +136,9 @@ export async function GET(
   const id = context.params.id;
   const searchParams = req.nextUrl.searchParams;
 
-  let system_keys: SystemKeys = {};
+  let system_keys: GFKeys = {};
   try {
-    system_keys = parse_system_keys(searchParams);
+    system_keys = parseGFKeys(searchParams);
   } catch (e) {
     console.error("error while parsing system keys:", e);
     // @ts-ignore
@@ -493,54 +492,4 @@ function parseSeedFromSearchParams({
   );
 
   return { seed, missing_required_hidden_fields };
-}
-
-interface SystemKeys {
-  [SYSTEM_GF_FINGERPRINT_VISITORID_KEY]?: string;
-  [SYSTEM_GF_CUSTOMER_UUID_KEY]?: string;
-  [SYSTEM_GF_CUSTOMER_EMAIL_KEY]?: string;
-}
-
-function parse_system_keys(
-  data: URLSearchParams | Map<string, string>
-): SystemKeys {
-  const map: SystemKeys = {};
-  const keys = Array.from(data.keys());
-  const system_gf_keys: string[] = keys.filter((key) =>
-    key.startsWith(SYSTEM_GF_KEY_STARTS_WITH)
-  );
-
-  for (const key of system_gf_keys) {
-    const value = data.get(key) as string;
-    switch (key) {
-      case SYSTEM_GF_FINGERPRINT_VISITORID_KEY: {
-        if (value.length === 32) {
-          map[key] = value;
-          break;
-        } else {
-          throw VISITORID_FORMAT_MISMATCH;
-        }
-      }
-      case SYSTEM_GF_CUSTOMER_UUID_KEY: {
-        if (is_uuid_v4(value)) {
-          map[key] = value;
-          break;
-        } else {
-          console.error("uuid format mismatch", value);
-          throw UUID_FORMAT_MISMATCH;
-        }
-      }
-      case SYSTEM_GF_CUSTOMER_EMAIL_KEY: {
-        if (!value.includes("@")) {
-          // TODO: more strict email validation
-          map[key] = value;
-          break;
-        }
-      }
-      default:
-        break;
-    }
-  }
-
-  return map;
 }
